+++ Terminal_Win_App/phase_2/serial_Win32_api/Test_serial/SerialTest2/SerialTest2/MyLib.h
@@ -15,7 +15,7 @@ typedef struct {
 } sData;
 
 void wait(unsigned timeout);
-uint8_t Reset_Buffer(uint8_t* array, uint8_t numb);
+uint8_t Reset_Buffer(uint8_t* array, int numb);
 uint8_t Initial_Preface_Of_Program(void);
 void    KeyBoard_Array_Input(char* Array, int number_element, char delimiter);
 uint16_t BDLib_Convert_Length_Char_to_Int_2(sData* Receive);
